fix PopularCommandExecutor test PopularCommandExecutor packages renamed
