--- src/main/java/edu/hw2/Main.java
@@ -1,14 +1,0 @@
-package edu.hw2;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public final class Main {
-    private final static Logger LOGGER = LogManager.getLogger();
-
-    private Main() {
-    }
-
-    public static void main(String[] args) {
-    }
-}
