+++ src/main/java/edu/hw2/task1/Expr.java
@@ -1,4 +1,4 @@
-package edu.hw2.Task1;
+package edu.hw2.task1;
 
 public sealed interface Expr {
     double evaluate();
