+++ src/main/java/edu/hw2/task2/Square.java
@@ -1,4 +1,4 @@
-package edu.hw2.Task2;
+package edu.hw2.task2;
 
 public class Square extends Rectangle {
     public Square(int sideLength) {
