+++ src/main/java/edu/hw2/task3/connection/Connection.java
@@ -1,4 +1,4 @@
-package edu.hw2.Task3.Connection;
+package edu.hw2.task3.connection;
 
 public interface Connection extends AutoCloseable {
     double PROBABLITY_BOUND = 0.5;
