+++ src/main/java/edu/hw2/task3/connectionmanager/ConnectionManager.java
@@ -1,6 +1,6 @@
-package edu.hw2.Task3.ConnectionManager;
+package edu.hw2.task3.connectionmanager;
 
-import edu.hw2.Task3.Connection.Connection;
+import edu.hw2.task3.connection.Connection;
 
 public interface ConnectionManager {
     double PROBABLITY_BOUND = 0.5;
