+++ src/main/java/edu/hw2/task3/connectionmanager/FaultyConnectionManager.java
@@ -1,7 +1,7 @@
-package edu.hw2.Task3.ConnectionManager;
+package edu.hw2.task3.connectionmanager;
 
-import edu.hw2.Task3.Connection.Connection;
-import edu.hw2.Task3.Connection.FaultyConnection;
+import edu.hw2.task3.connection.Connection;
+import edu.hw2.task3.connection.FaultyConnection;
 
 public class FaultyConnectionManager implements ConnectionManager {
     @Override
